% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RpackageFunctions.r
\name{convert_to_DNA}
\alias{convert_to_DNA}
\title{Convert SIRC results to DNAStringSet}
\usage{
convert_to_DNA(results, selection = "Cassettes", path_to_genome)
}
\arguments{
\item{results}{R list of results obtained via \code{\link{SIRCFinder}} function.}

\item{selection}{Type of sequences to extract:
\itemize{
  \item \code{"Cassettes"} - full SIRC sequences (default)
  \item \code{"DR"} - direct repeat elements only
  \item \code{"Spacers"} - spacer elements only
}
When selecting \code{"DR"} or \code{"Spacers"}, sequence IDs will include
additional suffix \code{";N"} where N is the element number within the cassette.}

\item{path_to_genome}{Path to the genome FASTA file.}
}
\value{
A \code{\link[Biostrings]{DNAStringSet}} object containing the requested sequences.
}
\description{
Extracts SIRC (Short Inverted Repeat Cassette) sequences from a genome FASTA file
and converts them into a \code{\link[Biostrings]{DNAStringSet}} object.
}
\details{
This function reads the genome FASTA file and extracts SIRC sequences based on
genomic coordinates provided in the \code{results} object.
}
\section{Sequence ID format}{

\itemize{
  \item For \code{selection = "Cassettes"}: \code{[chromosome]_[cassette_id]}
  \item For \code{selection = "DR"} or \code{"Spacers"}: \code{[chromosome]_[cassette_id];[element_number]}
}
}

\examples{
\dontrun{
# Assuming you have SIRC results and a genome file
sirc_results <- SIRCFinder(...)

# Extract full cassettes
cassettes <- convert_to_DNA(sirc_results,
                            selection = "Cassettes",
                            path_to_genome = "genome.fasta")

# Extract only direct repeats
drs <- convert_to_DNA(sirc_results,
                      selection = "DR",
                      path_to_genome = "genome.fasta")
}

}
\references{
Reference to be added after manuscript submission to bioRxiv.
}
\author{
Igor V Gorbenko \email{gorbenko@sifibr.irk.ru}
Karolina M. Zverintseva \email{zverincewa@mail.ru}
}
